Moved mob related events to their own package
