+++ src/com/palmergames/bukkit/towny/event/mobs/MobSpawnRemovalEvent.java
@@ -1,4 +1,4 @@
-package com.palmergames.bukkit.towny.event;
+package com.palmergames.bukkit.towny.event.mobs;
 
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Entity;
@@ -7,17 +7,17 @@ import org.bukkit.event.Cancellable;
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
 
-public class MobRemovalEvent extends Event implements Cancellable {
+public class MobSpawnRemovalEvent extends Event implements Cancellable {
 	private static final HandlerList handlers = new HandlerList();
 
 	private boolean cancelled = false;
 	private final Entity entity;
-	
-	public MobRemovalEvent(Entity entity) {
+
+	public MobSpawnRemovalEvent(Entity entity) {
 		super(!Bukkit.getServer().isPrimaryThread());
 		this.entity = entity;
 	}
-	
+
 	@Override
 	public HandlerList getHandlers() {
 		return handlers;
@@ -26,7 +26,7 @@ public class MobRemovalEvent extends Event implements Cancellable {
 	public static HandlerList getHandlerList() {
 		return handlers;
 	}
-	
+
 	@Override
 	public boolean isCancelled() {
 		return cancelled;
@@ -36,11 +36,11 @@ public class MobRemovalEvent extends Event implements Cancellable {
 	public void setCancelled(boolean isCancelled) {
 		cancelled = isCancelled;
 	}
-	
+
 	public Entity getEntity() {
 		return this.entity;
 	}
-	
+
 	public EntityType getEntityType() {
 		return this.entity.getType();
 	}
